Create a Harvest.id fixture in conftest.py
